refactor(cqrs): move composite manager package
