+++ pkg/cqrs/manager/cqrs.go
@@ -1,4 +1,4 @@
-package base_cqrs
+package manager
 
 import (
 	"cmp"
@@ -55,12 +55,17 @@ type adapterWithHelpers interface {
 	Helpers() driverhelp.DialectHelpers
 }
 
-func NewCQRS(adapter adapterWithHelpers) cqrs.Manager {
+func New(adapter dbpkg.Adapter) cqrs.Manager {
+	helperAdapter, ok := adapter.(adapterWithHelpers)
+	if !ok {
+		panic("bug: manager adapter does not implement dialect helpers")
+	}
+
 	// Force goqu to use prepared statements for consistency with sqlc
 	sq.SetDefaultPrepared(true)
 	return wrapper{
-		adapter: adapter,
-		q:       adapter.Q(),
+		adapter: helperAdapter,
+		q:       helperAdapter.Q(),
 	}
 }
 
@@ -69,6 +74,11 @@ type wrapper struct {
 	q       dbpkg.Querier
 	tx      *sql.Tx
 }
+
+var (
+	_ cqrs.Manager   = wrapper{}
+	_ cqrs.TxManager = wrapper{}
+)
 
 func (w wrapper) dialect() string {
 	return w.adapter.Helpers().GoquDialect()
