+++ pkg/cqrs/manager/history_queries.go
@@ -1,4 +1,4 @@
-package base_cqrs
+package manager
 
 import (
 	"context"
