+++ pkg/cqrs/manager/sqlite_converter.go
@@ -1,4 +1,4 @@
-package base_cqrs
+package manager
 
 import (
 	"encoding/json"
