+++ pkg/cqrs/manager/sqlite_converter_test.go
@@ -1,4 +1,4 @@
-package base_cqrs
+package manager
 
 import (
 	"database/sql"
